refactor: #52 requeest와 response 패키지 분리가 안 된 곳 수정 (#55)
