+++ src/main/java/com/backend/blooming/authentication/presentation/AuthenticationController.java
@@ -8,11 +8,11 @@ import com.backend.blooming.authentication.application.dto.TokenDto;
 import com.backend.blooming.authentication.infrastructure.oauth.OAuthType;
 import com.backend.blooming.authentication.presentation.anotaion.Authenticated;
 import com.backend.blooming.authentication.presentation.argumentresolver.AuthenticatedUser;
-import com.backend.blooming.authentication.presentation.dto.LogoutRequest;
-import com.backend.blooming.authentication.presentation.dto.WithdrawRequest;
+import com.backend.blooming.authentication.presentation.dto.request.LogoutRequest;
+import com.backend.blooming.authentication.presentation.dto.request.WithdrawRequest;
 import com.backend.blooming.authentication.presentation.dto.request.ReissueAccessTokenRequest;
 import com.backend.blooming.authentication.presentation.dto.response.LoginInformationResponse;
-import com.backend.blooming.authentication.presentation.dto.response.SocialLoginRequest;
+import com.backend.blooming.authentication.presentation.dto.request.SocialLoginRequest;
 import com.backend.blooming.authentication.presentation.dto.response.TokenResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
