--- src/main/java/com/backend/blooming/authentication/presentation/dto/WithdrawRequest.java
@@ -1,4 +1,0 @@
-package com.backend.blooming.authentication.presentation.dto;
-
-public record WithdrawRequest(String refreshToken) {
-}
