+++ src/main/java/com/backend/blooming/authentication/presentation/dto/request/LogoutRequest.java
@@ -1,4 +1,4 @@
-package com.backend.blooming.authentication.presentation.dto;
+package com.backend.blooming.authentication.presentation.dto.request;
 
 public record LogoutRequest(String refreshToken, String deviceToken) {
 }
