+++ src/main/java/com/backend/blooming/authentication/presentation/dto/request/SocialLoginRequest.java
@@ -1,4 +1,4 @@
-package com.backend.blooming.authentication.presentation.dto.response;
+package com.backend.blooming.authentication.presentation.dto.request;
 
 public record SocialLoginRequest(String accessToken, String deviceToken) {
 }
