+++ src/main/java/com/backend/blooming/authentication/presentation/dto/request/WithdrawRequest.java
@@ -1,0 +1,4 @@
+package com.backend.blooming.authentication.presentation.dto.request;
+
+public record WithdrawRequest(String refreshToken) {
+}
